delete context because there was an issue Relates #3
